+++ agent/agent_windows.go
@@ -1,0 +1,9 @@
+// +build windows
+
+package agent
+
+import "os"
+
+func watchForFlushSignal(flushRequested chan os.Signal) {
+	// not implemented
+}
